+++ veedrive/config.py
@@ -5,7 +5,12 @@ import pytz
 DEFAULT_HOST = os.getenv("VEEDRIVE_DEFAULT_HOST", "0.0.0.0")
 DEFAULT_PORT = os.getenv("VEEDRIVE_DEFAULT_PORT", 4444)
 
-SANDBOX_PATH = os.path.normpath(os.getenv("VEEDRIVE_MEDIA_PATH", "~/"))
+default_sandbox_folder = os.path.abspath(
+    os.path.join(os.path.dirname(__file__), "..", "tests", "sandbox_folder")
+)
+SANDBOX_PATH = os.path.normpath(
+    os.getenv("VEEDRIVE_MEDIA_PATH", default_sandbox_folder)
+)
 THUMBNAIL_CACHE_PATH = os.path.join(SANDBOX_PATH, "cache")
 
 STATIC_CONTENT_URL = os.getenv(
@@ -14,7 +19,6 @@ STATIC_CONTENT_URL = os.getenv(
 CONTENT_URL = os.getenv(
     "VEEDRIVE_CONTENT_URL", f"http://{DEFAULT_HOST}:{DEFAULT_PORT}/content"
 )
-DB_TYPE = os.getenv("VEEDRIVE_DB_TYPE", "mongo")
 DB_NAME = os.getenv("VEEDRIVE_DB_NAME", "test_database")
 DB_HOST = os.getenv("VEEDRIVE_DB_HOST", "127.0.0.1")
 DB_USERNAME = os.getenv("VEEDRIVE_DB_USERNAME")