+++ veedrive/content/test/test_unit.py
@@ -10,11 +10,7 @@ from .. import content_manager, fs_manager
 
 
 class TestPath(unittest.TestCase):
-    def setUp(self):
-        self.sandboxpath = config.SANDBOX_PATH  # '/tests/sandbox_folder/'
-
     def test_exceptions(self):
-
         with self.assertRaises(PermissionError):
             fs_manager.validate_path("/tmp", "dir")
         with self.assertRaises(PermissionError):
@@ -25,6 +21,7 @@ class TestPath(unittest.TestCase):
             fs_manager.validate_path(os.path.join(config.SANDBOX_PATH, "../"), "file")
         with self.assertRaises(PermissionError):
             fs_manager.validate_path(os.path.join(config.SANDBOX_PATH, "/"), "file")
+
         with self.assertRaises(FileNotFoundError):
             fs_manager.validate_path(
                 os.path.join(config.SANDBOX_PATH, "folder1", "file_not_exisiting"),
@@ -41,9 +38,6 @@ class TestPath(unittest.TestCase):
 
 
 class TestWsHandlers(unittest.TestCase):
-    def setUp(self):
-        self.sandboxpath = config.SANDBOX_PATH  # '/tests/sandbox_folder/'
-
     def test_create_file_response(self):
         path = "chess.jpg"
         obj = content_manager._create_file_url_response(path)
@@ -59,7 +53,6 @@ class TestThumbnail(unittest.TestCase):
 
 class TestContentOptimization(unittest.TestCase):
     cache_folder = "/tmp/testoptimized"
-    sandbox_path = config.SANDBOX_PATH
 
     def setUp(self):
         if os.path.exists(self.cache_folder):
@@ -68,11 +61,11 @@ class TestContentOptimization(unittest.TestCase):
     def test_portait_image(self):
         file_name = "NeuronGroup.jpg"
         img = cv2.imread(
-            os.path.join(self.sandbox_path, file_name), cv2.IMREAD_UNCHANGED
+            os.path.join(config.SANDBOX_PATH, file_name), cv2.IMREAD_UNCHANGED
         )
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 4096, 4096
+            file_name, config.SANDBOX_PATH, self.cache_folder, 4096, 4096
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
 
@@ -86,7 +79,7 @@ class TestContentOptimization(unittest.TestCase):
         os.remove(os.path.join(self.cache_folder, file_name))
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 4096, 512
+            file_name, config.SANDBOX_PATH, self.cache_folder, 4096, 512
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
         self.check_aspect(img, optimized)
@@ -98,7 +91,7 @@ class TestContentOptimization(unittest.TestCase):
         os.remove(os.path.join(self.cache_folder, file_name))
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 120, 512
+            file_name, config.SANDBOX_PATH, self.cache_folder, 120, 512
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
         self.check_aspect(img, optimized)
@@ -110,11 +103,11 @@ class TestContentOptimization(unittest.TestCase):
     def test_landscape_image(self):
         file_name = "HorizontalGroup.jpg"
         img = cv2.imread(
-            os.path.join(self.sandbox_path, file_name), cv2.IMREAD_UNCHANGED
+            os.path.join(config.SANDBOX_PATH, file_name), cv2.IMREAD_UNCHANGED
         )
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 4096, 4096
+            file_name, config.SANDBOX_PATH, self.cache_folder, 4096, 4096
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
         self.check_aspect(img, optimized)
@@ -126,7 +119,7 @@ class TestContentOptimization(unittest.TestCase):
         os.remove(os.path.join(self.cache_folder, file_name))
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 4096, 512
+            file_name, config.SANDBOX_PATH, self.cache_folder, 4096, 512
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
         self.check_aspect(img, optimized)
@@ -138,7 +131,7 @@ class TestContentOptimization(unittest.TestCase):
         os.remove(os.path.join(self.cache_folder, file_name))
 
         optimized = content_manager.optimize_image(
-            file_name, self.sandbox_path, self.cache_folder, 512, 120
+            file_name, config.SANDBOX_PATH, self.cache_folder, 512, 120
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
         self.check_aspect(img, optimized, float_precision=1)
@@ -149,10 +142,10 @@ class TestContentOptimization(unittest.TestCase):
 
     def test_image_with_size_smaller_than_box(self):
         path = "chess.jpg"
-        img = cv2.imread(os.path.join(self.sandbox_path, path), cv2.IMREAD_UNCHANGED)
+        img = cv2.imread(os.path.join(config.SANDBOX_PATH, path), cv2.IMREAD_UNCHANGED)
 
         optimized = content_manager.optimize_image(
-            path, self.sandbox_path, self.cache_folder, 4096, 4096
+            path, config.SANDBOX_PATH, self.cache_folder, 4096, 4096
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
 
@@ -161,9 +154,9 @@ class TestContentOptimization(unittest.TestCase):
 
         os.remove(os.path.join(self.cache_folder, path))
 
-        img = cv2.imread(os.path.join(self.sandbox_path, path), cv2.IMREAD_UNCHANGED)
+        img = cv2.imread(os.path.join(config.SANDBOX_PATH, path), cv2.IMREAD_UNCHANGED)
         optimized = content_manager.optimize_image(
-            path, self.sandbox_path, self.cache_folder, 4096, 500
+            path, config.SANDBOX_PATH, self.cache_folder, 4096, 500
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)
 
@@ -172,10 +165,10 @@ class TestContentOptimization(unittest.TestCase):
 
     def test_square_image(self):
         path = "chess.jpg"
-        img = cv2.imread(os.path.join(self.sandbox_path, path), cv2.IMREAD_UNCHANGED)
+        img = cv2.imread(os.path.join(config.SANDBOX_PATH, path), cv2.IMREAD_UNCHANGED)
 
         optimized = content_manager.optimize_image(
-            path, self.sandbox_path, self.cache_folder, 512, 512
+            path, config.SANDBOX_PATH, self.cache_folder, 512, 512
         )
         optimized = cv2.imread(str(optimized), cv2.IMREAD_UNCHANGED)